finish

+++ reset_dataset.py
@@ -5,7 +5,7 @@ from scipy.io import savemat
 import shutil
 def load_and_normalize_image(path, size=None):
     img = Image.open(path).convert('L')  # 灰度图
-    return np.array(img, dtype=np.int8)
+    return np.array(img)
 
 def create_right_aligned_stacks(folder_path, output_folder, stack_size=5):
     os.makedirs(output_folder, exist_ok=True)
@@ -36,7 +36,6 @@ def create_right_aligned_stacks(folder_path, output_folder, stack_size=5):
             else:
                 img = all_images[j]
             stack.append(img)
-
         stack_array = np.stack(stack, axis=0)  
         save_path = os.path.join(output_folder, os.path.basename(image_files[i]).replace('.png','.mat'))
         savemat(save_path, {'data': stack_array})
